requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0
httpx==0.27.0
tenacity==8.2.3
python-json-logger==2.0.7
pytest==8.0.2
pytest-asyncio==0.23.5
python-dotenv==1.0.1
alembic==1.13.1